Login

+++ data/src/main/java/com/github/sidky/data/paging/AbstractFirstPageLoader.kt
@@ -3,6 +3,7 @@ package com.github.sidky.data.paging
 import android.content.Context
 import androidx.work.Worker
 import androidx.work.WorkerParameters
+import com.github.sidky.data.apollo.TokenProvider
 import com.github.sidky.data.converter.ApolloPhotoListToDBConverter
 import com.github.sidky.data.converter.WorkResultUtil
 import com.github.sidky.data.dao.LoadingState
@@ -10,7 +11,6 @@ import com.github.sidky.data.dao.PhotoDAO
 import kotlinx.coroutines.runBlocking
 import org.koin.standalone.KoinComponent
 import org.koin.standalone.inject
-import timber.log.Timber
 
 abstract class AbstractFirstPageLoader(
     context: Context,
@@ -19,28 +19,39 @@ abstract class AbstractFirstPageLoader(
     private val converter: ApolloPhotoListToDBConverter by inject()
     private val dao: PhotoDAO by inject()
     private val loadingState: LoadingState by inject()
+    private val tokenProvider: TokenProvider by inject()
 
-    abstract suspend fun load(): PhotoLoaderResponse
+    abstract suspend fun load(): GraphQLResponse
 
     override fun doWork(): Result {
         return runBlocking {
-            Timber.e("BEGIN")
             val response = load()
-            val dbPhotos = converter.convertNullable(response.photos)
+            when (response) {
+                is GraphQLResponse.Success -> {
+                    val resp = response.response
+                    val dbPhotos = converter.convertNullable(resp.photos)
 
-            Timber.e("LOADED: ${dbPhotos}")
+                    val pagination = resp.next
 
-            val pagination = response.next
+                    loadingState.hasNext = pagination?.hasNext() ?: false
+                    loadingState.next = pagination?.next() ?: 0
 
-            loadingState.hasNext = pagination?.hasNext() ?: false
-            loadingState.next = pagination?.next() ?: 0
-
-            if (dbPhotos != null) {
-                dao.setPhotsWithURLs(dbPhotos)
-            } else {
-                dao.deleteAll()
+                    if (dbPhotos != null) {
+                        dao.setPhotsWithURLs(dbPhotos)
+                    } else {
+                        dao.deleteAll()
+                    }
+                    Result.success(WorkResultUtil.success(pagination))
+                }
+                is GraphQLResponse.Failure -> {
+                    refreshToken()
+                    Result.retry()
+                }
             }
-            Result.success(WorkResultUtil.success(pagination))
         }
+    }
+
+    suspend fun refreshToken() {
+        tokenProvider.refresh()
     }
 }
